% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{line_task_notify}
\alias{line_task_notify}
\title{Task end notification}
\usage{
line_task_notify(LINE_API_TOKEN = NULL, msg = "Task is finished.")
}
\arguments{
\item{LINE_API_TOKEN}{Your valid LINE Notify api token. Recomennded to save .Renviron file.}

\item{msg}{Message to show. 1,000 characters max. Defaults to "Task is finished."}
}
\description{
This function sends a notification to your LINE account via RStudio interface.
}
\seealso{
\url{https://notify-bot.line.me/doc/en/}
}
\author{
Koki Ando \email{koki.25.ando@gmail.com}
}
