% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineR.R
\docType{package}
\name{lineR}
\alias{lineR}
\alias{lineR-package}
\title{lineR: Sends line notification}
\description{
lineR package offers a function to send Line notifications using LINE Notify API <https://notify-bot.line.me/doc/en/>.
}
\seealso{
\url{https://notify-bot.line.me/doc/en/}
}
\author{
Koki Ando \email{koki.25.ando@gmail.com}
}
